requirements.txt
fastapi==0.70.1
uvicorn[standard]==0.32.0
python-multipart==0.0.7
python-dotenv==0.15.0
authlib==0.15.2
six==1.16.0
image_resizer @ git+https://github.com/wiating-app/image_resizer.git
auth0-python==3.13.0
boto3==1.16.18
elasticsearch==8.9.0
redis==3.5.3
aiofiles==22.1.0
python-keycloak==4.7.0
